+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.2.1] - 2022-09-13
+
+### Changed
+
+- Format DMB in memory before converting.
+- No longer skips models that lack textures.
+
+### Fixed
+
+- Formatted DMB now properly finds all entries, including those previously missed due to empty lines within entries.
+
 ## [0.2.0] - 2022-09-13
 
 ### Added
@@ -21,5 +32,6 @@
 - GUI app written in Python using the PySimpleGUI package.
 - Ability to convert `descr_model_battle.txt` from _Rome: Total War_ (OG) to _Total War: Rome Remastered_ (RR).
 
+[0.2.1]: https://gitlab.com/eb-online/tools/og2rr_dmb/-/releases/v0.2.1
 [0.2.0]: https://gitlab.com/eb-online/tools/og2rr_dmb/-/releases/v0.2.0
 [0.1]: https://gitlab.com/eb-online/tools/og2rr_dmb/-/releases/v0.1